requirements.txt
torch
torchvision
timm
facenet-pytorch
streamlit
Pillow
opencv-python-headless
numpy